Added task staggering, corrected Order API for order status updates, corrected Fronted to include Shipped status.

+++ Xango.Services.Queue.Processor/OrdersCancelledProcessor.cs
@@ -1,6 +1,8 @@
 ﻿using Microsoft.Extensions.DependencyInjection;
 using System;
 using System.Threading;
+using Xango.Models.Dto;
+using Xango.Services.Client.Utility;
 using Xango.Services.RabbitMQ.Utility;
 using static Xango.Services.Queue.Processor.QueueMessage;
 
@@ -19,8 +21,17 @@ namespace Xango.Services.Queue.Processor
 			if (this.AuthClient != null && this.AuthToken != null && this.OrderClient != null)
 			{
 				var orderHeader = message.OrderHeader;
+				Console.WriteLine($"[{this.GetType().FullName}] Processing order with ID {orderHeader.OrderHeaderId}.");
 				try
 				{
+					var correspondingOrderHeader = DtoConverter.ToDto<OrderHeaderDto>(this.OrderClient.GetOrder(orderHeader.OrderHeaderId).Result);
+					if (correspondingOrderHeader == null || correspondingOrderHeader.Status != SD.Status_Approved)
+					{
+						Console.WriteLine($"[{this.GetType().FullName}] Unable to retrieve order with ID {orderHeader.OrderHeaderId} and status Cancelled.");
+						RemoveOrderMessage();
+						return true;
+					}
+
 					var response = this.OrderClient.DeleteOrder(orderHeader.OrderHeaderId).Result;
 					if (response != null && response.IsSuccess)
 					{
@@ -32,6 +43,16 @@ namespace Xango.Services.Queue.Processor
 					Console.WriteLine($"[{this.GetType().FullName}] Exception: {exc.Message}");
 				}
 			}
+			if (processed)
+			{
+				Console.WriteLine($"[{this.GetType().FullName}] Successfully processed order with ID {message.OrderHeader.OrderHeaderId}.");
+				RemoveOrderMessage();
+			}
+			else
+			{
+				Console.WriteLine($"[{this.GetType().FullName}] Failed to process order with ID {message.OrderHeader.OrderHeaderId}.");
+			}
+			return processed;
 
 			return processed;
 		}
